project: remove support for -m option
In order to add support for extension commands, we will need to know
where the manifest is before we parse any command line arguments.

This means the -m argument supported by the existing project commands
has to go away. We'll always let the manifest module fetch the
contents from the location specified in the configuration file.

(The Manifest class is still capable of parsing manifests from
arbitrary file system paths.)

While we're here, add a validity check for the sections parameter to
the Manifest constructor and factory methods. This wasn't being used
properly: it should be a list. The fact that strings are sequences
means it basically gets silently ignored when passing a string.

Signed-off-by: Marti Bolivar <marti@foundries.io>