west: add infrastructure for external commands
Add support for a new optional 'west-commands' key in each project in
a manifest. This specifies the location of a file which describes
additional west commands provided by that project.

Extend the Manifest class to detect this key and record the
information inside. Nothing else is done about external commands in
that class, since its purpose is just to parse the manifest.

Extend west.commands with a new external_commands() function, which
consumes a Manifest and returns a list of "descriptors" of external
commands the manifest's projects declare. These descriptors contain
metadata about the commands, but do not actually import any Python
modules related to them or otherwise try to instantiate the relevant
WestCommand instances.

This is for two reasons:

1. Performance: don't import what you might not need

2. Security: don't import random code you downloaded somewhere on the
   internet! Importing a module you got from somewhere random is
   basically curl | sh.

This infrastructure will be wired into main() in a subsequent patch.

Signed-off-by: Marti Bolivar <marti@foundries.io>