commands: allow extension commands to provide help strings
For various reasons, we don't load extension command modules until the
user has requested that we run a particular extension command.

As is, this limits the "west -h" output to the extension command names
and the projects which define them, which isn't as nice as what
built-in commands get.

Allow extension commands to provide help text by extending the schema
for each command entry with an optional "help" key, and display it in
the output. If it is missing, a default is provided which recommends
just running "west $EXTENSION_COMMAND_NAME -h".

Fix the help formatting by adding a colon after each extension command
name to match the style used for built-ins while we are here.

Keep test cases up to date.

Signed-off-by: Marti Bolivar <marti@foundries.io>