commands: fix for naming invalid project from project directory
The behavior of 'west list SOME_INVALID_PROJECT' (and any other
commands which try to match SOME_INVALID_PROJECT with real project)
is incorrect when called from a project directory: the current
project is found instead of an error being returned.

Fix this by requiring that, when treated as a path, the given argument
exactly normalizes to a project directory. Otherwise, it is possible
for it to match nonexistent files inside of the existing project
directory.

Add some more testing. The ones that expect a CalledProcessError fail
without this project.py patch. With it, they all pass.

Signed-off-by: Marti Bolivar <marti@foundries.io>