bootstrap: Removed second fetch and allow for SHA
This commit cleans up the second fetch, which potentially could have
changed content since first fetch.

Normal ref-space (similar to git clone) and user specified revision is
now fetched in same command.

This further allows to checkout tags in detached HEAD while still
referring to the tag name instead of FETCH_HEAD.

This commit also allows a user to specify a sha.
When specify a sha as revision, that sha must be valid in to the normal
ref-space.

Signed-off-by: Torsten Rasmussen <torsten.rasmussen@nordicsemi.no>