bootstrapper: improve heuristic for --zephyr-base
We're seeing warnings when people have ZEPHYR_BASE set to the
"<directory>" they pass to "west init -l <directory>".

Avoid them exactly in the situation that <directory> is a zephyr
repository by using the manifest directory as ZEPHYR_BASE. For
upstream, that is correct anyway.

Merge a couple of lines while here as a cosmetic improvement.

Fixes: #167

Signed-off-by: Marti Bolivar <marti@foundries.io>