commands: fine-grained extension command error handling
Rename BadExternalCommand to ExtensionCommandError and make it a
subclass of CommandError so that it fits naturally in the hierarchy
and provides a returncode.

Add a hint attribute to it so that we can print further diagnostic
information to the user when things go wrong.

While we are here, use traceback.print_exc() in the error handlers at
main level to make it more clear that the verbosity is not altering
the control flow of the program, just what is being printed. Also use
the returncode attribute in a CalledProcessError since that is
available.

Print the entire command invocation that caused the problem with -v
added instead of making the user guess what to do.

Signed-off-by: Marti Bolivar <marti@foundries.io>