zephyr-base: ZEPHYR_BASE env takes precedence over config settings
This commit changes the order of precedence so that environment setting
takes precedence over west config option.
It also introduces a new config file setting, to allow to specify which
setting: environment or configfile that will take precedence.
To simplify search for ZEPHYR_BASE, this setting will now be written to
the local .west/config during init and selfupdate, if non-existing.

Signed-off-by: Torsten Rasmussen <torsten.rasmussen@nordicsemi.no>