More clean up for manifest.py / project.py responsibilities
This commit is further progress on #191, but doesn't fix it entirely.

West's design has always been to split code into reusable libraries
and commands which wrap them. The extra effort is justified since West
is pluggable by extension commands, which often benefit from being
able to use the library pieces.

Although the code in project.py started out self-contained, it's
gradually being refactored to fit the design.

Continue that by factoring two more features into manifest.py: helpers
for running git commands in a project and formatting strings using
project attributes.

This lets us remove almost all of the wrappers around west.log in a
way that makes the code easier to read. The remaining _inf() is
renamed _banner() to make its purpose clearer.

Beyond being a useful refactoring which makes extensions easier to
write, this is needed to fix a bug where west doesn't check if the
currently running version is not a descendant of the manifest
revision, indicating it is out of date and prompting a warning. Fixing
that is a high priority issue for Zephyr v1.14.

Delete a fair bit of dead code left over from the older multirepo
commands which were deleted from v0.4.0 while we are here.

Signed-off-by: Marti Bolivar <marti@foundries.io>