commands: fix configuration option to disable extensions
We need to be using getboolean() to parse "true" and "false" strings into their
equivalent booleans. This isn't happening, so setting the option to any
nonempty string is continuing to allow extensions to run. Fix that, and move
the config option parsing to extension_commands() to make sure the fix only
needs to happen once.

Fixes: #238
Signed-off-by: Marti Bolivar <marti@foundries.io>