manifest: error if self.path doesn't match directory name
As an example, the "Clone the Zephyr Repositories" section of Zephyr's
Getting Started Guide provides the following command to wrap a west
installation around an already cloned, standalone zephyr repository:

  west init -l zephyr/

It feels natural to use the same command with some alternative
zephyr-blue/ or zephyr-red/ directory name, however that's not
compatible with the zephyr-blue/west.wml manifest which has its
self.path hardcoded to "zephyr". Hardcoding self.path is not a issue,
the problem is that the inconsistency is not caught and west pretends
the init command was successful. It's then much more difficult to
understand later why the build fails to find some file(s).

Implement a sanity check which make west init "fail fast" like this:

  FATAL ERROR: Manifest self.path is 'zephyr' but located in \
               directory '/home/marc/work/zephyr-blue'

Fixes #267

Signed-off-by: Marc Herbert <marc.herbert@intel.com>