commands: config: try to improve -h output
Align the description with the west.configuration docstring.

Separate the options which set the configuration file to use into
their own named argument group in the help output. List them in the
same system, global, local order that they appear in the description.

This requires using add_argument_group() instead of
add_mutually_exclusive_group(), as the latter can't take a group
name. Preserve mutual exclusion with a custom action with equivalent
semantics. This changes part of the usage from:

   [--system | --global | --local ]

to:

   [--system] [--global] [--local]

That's a bit unfortunate, but we're about to add other options that
don't belong in this group, and I think it's worth it to have these
three show up on their own.

While we're here, don't make the help output depend on the
platform. Let's keep it uniform.

Signed-off-by: Marti Bolivar <marti.bolivar@nordicsemi.no>