manifest: allow Project.git to capture stderr
It's not always possible to silence git commands with -q, and it is
sometimes necessary to do so to avoid spurious output being shown to
the user.

One example is using "git cat-file -e SOME_REV" to test if a revision
has been fetched and is available in the local object database. This
prints to stderr if it isn't. The command or other code running
Project.git() in this case may be able to detect and deal with the
missing revision appropriately without wanting the user to see any
messages, and that's currently not possible.

Fix it by adding a capture_stderr that works the same way
capture_stdout does, but for stderr.

Signed-off-by: Marti Bolivar <marti.bolivar@nordicsemi.no>