commands: save the parsed manifest from main
This allows individual commands to avoid re-parsing. Properly
initialize it in the constructor (and topdir while we're here).

Use a property to get and set the manifest to allow us to fail with a
consistent error message when commands that require a manifest don't
get one.

We are basically saving the manifest similarly to topdir.

This opens up opportunities to avoid re-parsing the manifest
repeatedly.

Signed-off-by: Marti Bolivar <marti.bolivar@nordicsemi.no>