commands: project: use cached manifest instead of reparsing
Now that we have a self.manifest attribute which returns the manifest
when it's available and fails gracefully when it isn't, we can avoid
re-parsing the manifest in the project command implementations.

Note this change is more convenient to make after having introduced
_ProjectCommand to allow accessing per-instance state from each of
these commands, instead of using standalone module-level functions.

Signed-off-by: Marti Bolivar <marti.bolivar@nordicsemi.no>