commands: do not require cloned projects in west list
That's unnecessary for almost all the format strings, and users may
want to get information about projects they don't have cloned in order
to decide whether or not to clone them.

Use the fact that we're delaying evaluation of all git related
commands to exit out only if the sha of an uncloned project is
requested. Otherwise, all the information we need is in the manifest.

Signed-off-by: Marti Bolivar <marti.bolivar@nordicsemi.no>