commands: project: consolidate error handling behavior
The current way we handle errors when a command means an operation has
to be done on multiple projects (such as "west diff", "west update",
etc.) is unfortunate and a bit inconsistent.

In particular, "west update" tries in some cases to keep going and
update as many projects as possible when a rebase fails, but "west
diff" will abort on the first failure. Even worse, "west update" calls
the _git() helper in various places, which means that those commands
failing halts the entire update, leaving *only* failed rebases to
allow the process to continue.

Other commands, like "west forall", don't capture errors at all, so
the failure of a subcommand does not result in the overall process
failing. Whoops.

Let's clean up this mess by adding a helper function,
_handle_failed(), to our shared parent class, which makes sure that if
any error occurred, the whole command fails. Use it from diff, status,
update, and forall.

Change the command implementations as necessary so they never call
_git(), which allows us to roll up errors from anywhere in the command
procedure without aborting the program immediately. This lets us
inline quite a bit of code while we're here.

The lone exceptions in this file are init, list, and selfupdate.
The first two still fail immediately on the first error; the final one
continues to scream and die.

I think at this point I'm willing to say the manifest.py vs project.py
responsibilities are cleaned up:

- west.manifest.Project's git helpers are for general purpose and nondestructive
  use.
- west.commands.project's use of Project and its git helpers are to do
  destructive operations, but also invoking the nondestructive ones
- multi-project git operations have to be handled outside of
  west.manifest, e.g. in _ProjectCommand

So:

Fixes: #191

Signed-off-by: Marti Bolivar <marti.bolivar@nordicsemi.no>