West v0.6.3
This fixes an issue observed with non-ascii content in CMake caches in
west.cmake. This module is still used by the Zephyr v1.14 LTS, so
we'll keep carrying fixes for it until that's no longer the case.

After v1.14 is history, both west.build and west.cmake will be removed.

Signed-off-by: Marti Bolivar <marti.bolivar@nordicsemi.no>