project: add import support to 'west update'
Adjust west init to ignore imports. It doesn't need anything other
than what's available in the top level manifest.

Add an importer callback from west update, which can handle
resolution when project data is needed but not available. Move the
_update() private helper to an instance method. This helps us keep
more state inside of self, which is useful now that we want to invoke
this method from an instance-specific callback.

This imposes some limitations when west update is used with manifest
imports. In particular, it is not permitted to 'west update
SOME_PROJECT' if the manifest uses imports.

The reason why is that it would require the west.manifest API to track
extra information during manifest resolution that it's not capable of
doing, and which would take more work to do than we think is
worthwhile right now.

One pathological example follows.

my-manifest/west.yml:

    manifest:
      projects:
        # 'my-upstream' is the upstream manifest that defines its
        # own projects. I want all of them. I want to track master.
        - name: my-upstream
          url: https://github.com/my-vendor/upstream-manifest
          import: true

Suppose my-upstream points at commit A right now, i.e. "git rev-parse
refs/heads/manifest-rev" prints "A", but the remote fetch URL
https://github.com/my-vendor/upstream-manifest has a master commit
pointing at B.

Further, suppose we have:

my-upstream/west.yml at commit A:

manifest:
  # ...
  projects:
    - name: p1

my-upstream/west.yml at commit B:

manifest:
  # ...
  projects:
    - name: p2

Think about what would happen if the user runs:

$ west update

In this case, my-upstream is updated, and therefore its manifest-rev
should now point at commit B. This introduces project p2, which is not
cloned on the local file system. So p2 should be fetched.

The p1 project is 'abandoned': it still exists on the local file
system, but west forgets about it entirely unless we do something like
set my-upstream's revision to point back at A.

OK, now what if we run:

$ west update p2

Given the above, this is a seemingly sensible request: if plain 'west
update' would clone and update p2, then 'west update p2' should also
work, no? Or should it?

If yes, that would mean west would need to speculatively update all
projects that have import statements, hunting for a p2 project. We
can't do that right now, and it's not clear how such a thing would be
defined.

Or what about this?

$ west update p1

If plain 'west update' means p1 is no longer a project, this command
should fail, no? Or should it succeed?

If it should fail, that would require west to fetch from
my-upstream (since that's the source of the manifest data that defined
p1 in the first place) and check if it's still there, and fail when
it's not.

If it should succeed, why would 'west update' with no arguments
produce entirely different results?

The presence of pathological examples like these is too hard to deal
with for an initial implementation, so we're going to intentionally
limit 'west update' to take no arguments when manifest imports are
used.

Let's see if people care, and if they do, how they would propose
resolving situations such as the one above.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>

Initial limitation is:

Suggested-by: Carles Cufí <carles.cufi@nordicsemi.no>