manifest: validate() fixes
We need to be a bit more careful with our types.

For example, the string "not-a-manifest" can be loaded as YAML, and
evaluates to... itself. And the string 'manifest' sure is in the
string 'not-a-manifest', so the "'manifest' not in data" check in
validate() doesn't catch the invalidity.

Then we try to index ['manifest'] in a string, which doesn't work,
because string indexes must be integers. Boom.

Fix it with extra type checking.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>