manifest: refactor git checks into project commands
It's nicer for the user of the manifest library (e.g. 'west diff') to
decide what should happen if git is not available, so move some
warning calls about missing git into project.py, and add a new helper
for dealing with dying when git is not found.

This lets us keep using log.wrn() for this detection, which we are
about to stop doing from manifest.py.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>