manifest: pass projects map down the tree
Pass the "known projects" map to the Manifest constructor via a kwarg.
This makes it possible to track most of the known projects when
resolving imports. That in turn partially resolves #360, except in the
case of import maps, where we are still creating a whole new manifest
and then filtering it out. We can fix that next.

This implies a hacky change to _check_paths_are_unique(): we need to
ignore checks involving the manifest repository if we're not top
level, as they may lead to false positives. In my opinion, this is yet
another good reason to avoid mixing the manifest repository itself
into the projects list.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>