manifest: pass import map filters down the tree
As is already done for projects, pass the filter function implied by
an import's whitelists and blacklists down the call stack. When we
encounter a new filter due to a map import, compose the existing
filter with the new one. We use None as a signal that there is no
filter in place to avoid calling unnecessary functions.

The motivation behind this patch was to correctly log project addition
exactly once, when they are first discovered in the recursion tree.

However, it's actually necessary for correctness, and fixes a bug that
can be shown by example:

west.yml:

  manifest:
    ...
    self:
      import:
        file: foo.yml
        name-whitelist: just-this-project

foo.yml:

  manifest:
    ...
    self:
      import: bar.yml

bar.yml:

  manifest:
    ...
    projects:
    - name: just-this-project
    - name: but-not-this-one

Above, we should ultimately import just-this-project from bar.yml, and
not but-not-this-one. This implies propagating and ANDing together
filters implied by whitelists and blacklists from higher levels in the
tree.

This patch also includes a test case for this type of situation.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>