west update: allow PROJECTs defined in manifest repo
Relax the restriction that west update cannot update individual
projects when manifest imports are in use, as long as the projects on
the command line are all defined in the manifest repository.

That's a well-defined enough operation to support, since this type of
project is basically equivalent to being defined in the top level
west.yml.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>