manifest: factor private context into a namedtuple
I want to add more values to the private context we pass around as we
resolve the manifest, but it's annoying to do so one parameter at a
time, since it requires updating a lot of functions.

Let's make a container object for this state and pass it around instead.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>